moved config type into models
