+++ backup-service/config/config.go
@@ -4,29 +4,18 @@ import (
 	"log/slog"
 	"os"
 
+	"github.com/Regncon/conorganizer/backup-service/models"
 	"github.com/joho/godotenv"
 )
 
-// Config Application configs and secrets
-type Config struct {
-	AWS_ENDPOINT_URL_S3   string
-	AWS_ACCESS_KEY_ID     string
-	AWS_SECRET_ACCESS_KEY string
-	AWS_REGION            string
-	BUCKET_NAME           string
-	DB_PREFIX             string
-	DISCORD_WEBHOOK_URL   string
-	DISCORD_SECRET_KEY    string
-}
-
 // Load Loads the required system environment variables
-func Load(logger *slog.Logger) Config {
+func Load(logger *slog.Logger) models.Config {
 	err := godotenv.Load()
 	if err != nil {
 		logger.Error("Error loading .env file")
 	}
 
-	return Config{
+	return models.Config{
 		AWS_ENDPOINT_URL_S3:   os.Getenv("AWS_ENDPOINT_URL_S3"),
 		AWS_ACCESS_KEY_ID:     os.Getenv("AWS_ACCESS_KEY_ID"),
 		AWS_SECRET_ACCESS_KEY: os.Getenv("AWS_SECRET_ACCESS_KEY"),
